+++ src/check/arbitrary/WebArbitrary.ts
@@ -4,7 +4,9 @@ import { constant } from '../../arbitrary/constant';
 import { buildAlphaNumericPercentArb } from './helpers/SpecificCharacterRange';
 import { domain, hostUserInfo } from './HostArbitrary';
 import { nat } from '../../arbitrary/nat';
-import { ipV4, ipV4Extended, ipV6 } from './IpArbitrary';
+import { ipV4 } from '../../arbitrary/ipV4';
+import { ipV4Extended } from '../../arbitrary/ipV4Extended';
+import { ipV6 } from '../../arbitrary/ipV6';
 import { oneof } from '../../arbitrary/oneof';
 import { option } from '../../arbitrary/option';
 import { stringOf } from '../../arbitrary/stringOf';
