+++ src/fast-check-default.ts
@@ -52,7 +52,9 @@ import { integer, IntegerConstraints } from './arbitrary/integer';
 import { maxSafeInteger } from './arbitrary/maxSafeInteger';
 import { maxSafeNat } from './arbitrary/maxSafeNat';
 import { nat, NatConstraints } from './arbitrary/nat';
-import { ipV4, ipV4Extended, ipV6 } from './check/arbitrary/IpArbitrary';
+import { ipV4 } from './arbitrary/ipV4';
+import { ipV4Extended } from './arbitrary/ipV4Extended';
+import { ipV6 } from './arbitrary/ipV6';
 import { letrec } from './arbitrary/letrec';
 import { lorem, LoremConstraints } from './arbitrary/lorem';
 import { mapToConstant } from './arbitrary/mapToConstant';
